% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-parsed-model.R
\name{as_parsed_model.modeldb_lm}
\alias{as_parsed_model.modeldb_lm}
\title{Prepares parsed model object}
\usage{
\method{as_parsed_model}{modeldb_lm}(x)
}
\arguments{
\item{x}{A parsed model object}
}
\description{
Prepares parsed model object
}
