% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidypredict.R
\name{as_parsed_model}
\alias{as_parsed_model}
\title{Create a parsed_model table}
\usage{
as_parsed_model(df, model_type = "lm")
}
\arguments{
\item{df}{A Local or remote data frame}

\item{model_type}{the value passed to the model type. Defaults to 'lm'}
}
\description{
Converts the output of linear_regression() to tidypredict readable output
}
\examples{
library(dplyr)

mtcars \%>\%
  select(mpg, wt) \%>\%
  linear_regression_db(mpg) \%>\%
  as_parsed_model()

}
