% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{linear_regression}
\alias{linear_regression}
\title{Linear Regression functions}
\usage{
linear_regression(df, y_var = NULL, sample_size = NULL)
}
\arguments{
\item{df}{A Local or remote data frame}

\item{y}{Dependent variable}

\item{x1}{First independent variable}

\item{x2}{Second independent variable}

\item{output}{The format of the output. Defaults to a 'tidypredict' parsed model}
}
\description{
It uses 'tidyeval' and 'dplyr' to to apply the a simple
linear regression formula with one or two independent
variables
}
\details{
It currently only works with continous dependent variables,
no support for categorical variables yet.
}
\examples{
library(dplyr)

mtcars \%>\%
  two_variable_regression(mpg, wt, qsec)

mtcars \%>\%
  simple_linear_regression(mpg, wt)

}
