% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeldb.R
\docType{package}
\name{modeldb-package}
\alias{modeldb}
\alias{modeldb-package}
\title{modeldb: Fix models inside the database}
\description{
Fix models inside the database
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/edgararuiz/modeldb}
  \item Report bugs at \url{https://github.com/edgararuiz/modeldb/issues}
}

}
\author{
\strong{Maintainer}: Edgar Ruiz \email{edgar@rstudio.com}

}
\keyword{internal}
