% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-parsed-model.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_parsed_model}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{tidypredict}{\code{\link[tidypredict]{as_parsed_model}}}
}}

