% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeldb.R
\docType{package}
\name{modeldb-package}
\alias{modeldb}
\alias{modeldb-package}
\title{modeldb: Fits Models Inside the Database}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Uses 'dplyr' and 'tidyeval' to fit statistical models inside the database. It currently supports KMeans and linear regression models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/modeldb}
  \item Report bugs at \url{https://github.com/tidymodels/modeldb/issues}
}

}
\author{
\strong{Maintainer}: Edgar Ruiz \email{edgar@rstudio.com}

}
\keyword{internal}
